Prometheus setup fixed + replaced poppy with max
Prometheus setup needed a re-run as it was somewhat borked on roy;
in doing so, went through trying to get things setup with podman.
Eventually abandoned containerized setup for prom because DNS resolution
failed (after much messing around with it not being able to write to my
bind mount). Resolution was to go back to Raspbian package. Also needed
to fix up part of the ansible setup to account for missing folders
on first installation.

Poppy has died and been replaced with max. Should be set up identically.